add useMarkers hook
